30,31: order-detail feature

+++ src/main/java/com/project/shopapp/dtos/OrderDetailDTO.java
@@ -21,7 +21,7 @@ public class OrderDetailDTO {
 
     @JsonProperty("number_of_products")
     @Min(value = 1, message = "Number of products must be > 0")
-    private Long numberOfProducts;
+    private Integer numberOfProducts;
 
     @Min(value = 0, message = "Price must be >= 0")
     private Float price;
